Backport of [27112] from trunk
Replaced uses of GetDecoder() with decoder (it's internal to the base class, 
no need to use the external method). Also changed myFramesRead so instead of 
counting locally, read decoder->GetFramesRead(), as the decoder is properly 
counting frames, and we are getting *fields* returned to us for H.264. Trying i
to count locally broke H.264 interlaced playback. This is a better way.

Fixes #9126 again, and without breaking playback this time.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-24-fixes@27113 7dbf422c-18fa-0310-86e9-fd20926502f2