Lock the OSD when the teletext decoder is updating the teletext screen and fix an OSD locking typo.
r27011 removed the code that paused the player/decoder while the OSD is
reinitialised (as it can cause lockups in conjunction with the letterbox
detection code - and I thought it was safe to do so!). The teletext
decoder however updates the teletext screen directly and hence it can
crash hard (N.B. the other caption screens do not operate in the same
way).

So lock the OSD while the update takes place (which will also reset the
teletext viewer).

Closes #9166


git-svn-id: http://svn.mythtv.org/svn/trunk@27114 7dbf422c-18fa-0310-86e9-fd20926502f2