Backport r27119 from trunk to release-0-24-fixes
Disable support for the Broadcom BCM70012 CrystalHD hardware decoding
card (the older Broadcom device).

I now have a BCM70012 to test with and first impressions are that the
code fixes to provide proper support are not as simple as I previously
thought. When it is fixed, I will no doubt backport to 
release-0-24-fixes


git-svn-id: http://svn.mythtv.org/svn/branches/release-0-24-fixes@27125 7dbf422c-18fa-0310-86e9-fd20926502f2