Simply the audio sync offset code.
This removes the ResetAudioTimecodeOffset method, which is superfluous,
and the ResyncAudioTimecodeOffset, which breaks av sync here. The
actions required to trigger these methods are both, as far as I can
tell, undocumented (and hence I doubt anyone is actually using them) and
it allows a degree of simplification of the code - most importantly
ensuring that only the player retains any offset value.

git-svn-id: http://svn.mythtv.org/svn/trunk@27147 7dbf422c-18fa-0310-86e9-fd20926502f2

Backports SHA:873ba864