Make the audio sync adjustment persistent.
This saves any manual av sync adjustment to the database and retrieves
any stored value at the start of playback without adding an additional
setting to the audio/playback settings.

The fix will not suit all - for example where an offset is required only
for certain audio codecs - but it is consistent with the current
approach and probably addresses the most common problem.

This additionally removes the savedAudioTimecodeOffset variable and
ensures the offset is always restored after a seek - which is one less
thing to worry about when the code is changed and fixes at least one
instance where the offset was not being restored. Calls to
ClearAfterSeek that were previously not saving the offset first were
either broken or the audio offset is not important (e.g. preview
generator).

git-svn-id: http://svn.mythtv.org/svn/trunk@27148 7dbf422c-18fa-0310-86e9-fd20926502f2

Backport: SHA:9dbe19fb