Remove the use of OpenGL immediate mode by using vertex arrays.
This moves us most of the way to supporting OpenGL ES 2.0 for painting
the UI but there is some compatability work required to use
glVertexAttribPointer for ES2.0 - which is part of the OpenGL 2.0
specification (and I don't think we're ready to bump the OpenGL
requirement to 2.0 just yet). The remainder of the 'new' functionality
we are using is OpenGL 1.0 - and anything that doesn't support that
probably isn't worth supporting.

The patch is complicated by the use of both QRect and QRectF in the
render code and could probably be simplified in the future.

This also opens up the possibility of using Vertex Buffer Objects for
vertex data but I can't honestly see the need at the moment. We probably
just aren't pushing enough vertex data to really see any performance
improvement.


git-svn-id: http://svn.mythtv.org/svn/trunk@27185 7dbf422c-18fa-0310-86e9-fd20926502f2