Add initial support for Vertex Buffer Objects.
This is the first of a series of commits aimed at 'profiling' our OpenGL
support. While I've not entirely settled on how this will work, the
broad idea is to use 3 levels of functionality:-

- Shiny. OpenGL 2.0 (and above) and OpenGL ES 2.0. Utilising Vertex
Buffer Objects and Pixel Buffer Objects for optimised updating of vertex
and image data respectively, GLSL for texturing, colourspace conversion,
deinterlacing and UI effects and FrameBuffer Objects for multi-pass
filtering.

- Legacy. Use Vertex Arrays, multi-texturing and fragment programs with
limited or no support for GPU based filters and effects.

- Fallback. Software based colourspace conversion.

Realistically level 3 should only be relevant to broken installations
(and I may yet decide to drop the software fallback) and the bulk of
current OpenGL setups should be perfectly happy with a full featured
approach.

Various other extensions currently used, such as YCbCr textures and
fences, will be used as available.


git-svn-id: http://svn.mythtv.org/svn/trunk@27223 7dbf422c-18fa-0310-86e9-fd20926502f2