Renumbered failure exits to be > 256
This is necessary as in at least one place (mythwelcome) we use the exit code
byte bitmapped.  The only way for that to work correctly is if we are not
mapping our errors into the same space.  As any users of myth_system should be
either testing for 0/non-zero, or looking for specific errors, this shouldn't
affect anything.