Make the Studio Levels 'filter' a persistent database setting.
- add a keybinding to toggle studio levels (default - no binding).
- add an OSD menu option to toggle studio levels when available.
- store the value in the database along with other persistent settings
such as volume, av sync offset, brightness, contrast, colour and hue.

This is currently only available with VDPAU playback and the old
vdpaustudio filter option is removed.

I've also allowed toggling studio levels via the TOGGLEPICCONTROLS
binding - though as it is a binary setting, it doesn't look sensible in
relation to the other playback controls that have a full range.


git-svn-id: http://svn.mythtv.org/svn/trunk@27322 7dbf422c-18fa-0310-86e9-fd20926502f2