Fix the in-recording seektable for MPEG2-TS.  Also, make the in-recording seektable for MPEG2-PS better, although not yet quite correct.  In particular, the file offsets are consistently earlier than they should be, but at least this makes the previews functional.
The original issue seemed to be that we were recording the offest AFTER the
packet in question, rather than the offset at the beginning of the packet.  In
MPEG2-TS, due to the packetizing of the stream, it's easy to backtrack to the
beginning of the previous packet, but in MPEG2-PS, the frames are all of 
differing sizes, and I haven't quite got it tweaked right yet.

Refs #9256



git-svn-id: http://svn.mythtv.org/svn/trunk@27338 7dbf422c-18fa-0310-86e9-fd20926502f2