Improved teletext display.
- fix some font stretch adjustments for double height text by basing it
on the actual stretch value.
- be a little more agressive in determining the font height.
- ignore the UI determined font stretch value and select the largest
possible value that will fill the width available for each character.

Collectively these ensure the font used is as large and as visible as
possible, especially on widescreen displays with VDPAU/OpenGL/Direct3D
and widescreen video sources (XVideo).

Use of a fixed space font is still highly recommended.


git-svn-id: http://svn.mythtv.org/svn/trunk@27367 7dbf422c-18fa-0310-86e9-fd20926502f2