Refactor TeletextDecoder and TeletextScreen and add a new TeletextReader class.
This follows the approach used in the other subtitle and caption classes
whereby the screen object handles OSD representation of the data, the
reader object holds the state and data and the decoder object feeds new
data into the reader.

This simplifies setup and initialisation of the teletext objects and
removes the requirement to lock the OSD from the decoder, which was
causing endless problems.

The previous behaviour of always decoding teletext packets is also
reinstated which radically improves responsiveness. Actually decoding
and parsing the packets is trivial for a modern processor and hence
should have no impact on performance.

I will backport to release-0-24-fixes in a few days if there are no
reported problems.

Refs #9271


git-svn-id: http://svn.mythtv.org/svn/trunk@27381 7dbf422c-18fa-0310-86e9-fd20926502f2