Refactor the RingBuffer classes.
-- make distclean STRONGLY recommended --

This creates a proper class hierarchy for the RingBuffer classes, adds a
new FileRingBuffer class and removes the capitalisation from the file
names.

As noted in #9287, this a straight refactor with no attempt made to
optimise the interface or minimise code duplication.

Tested on 3 different boxes here with no apparent regressions.

Refs #9287


git-svn-id: http://svn.mythtv.org/svn/trunk@27413 7dbf422c-18fa-0310-86e9-fd20926502f2