Tweak PWD in the Makefile to be more specific
If the Makefile is started with "make -C mythtv" from a Makefile another level
up, $PWD is still seeming to be set from the top level Makefile.  This forces
a reset of $PWD from the shell at this level, causing version.sh to put
version.cpp in the correct place.