Fix file streaming to stay working past 4G
The previous logic would loop, and remove 2G from llBytes the first time
through the loop (correctly), but remove 4G from llBytes the second time
through the loop (very incorrectly).  This causes a UPnP client, etc to
abruptly stop after 4G if the file size is between 4G and 6G, and 2G early for
file size of 6G to 8G, and so on.

I simplified the logic to work correctly in my testing.