Update Vimeo and Youtube scripts/HTML to allow remote-based control.
Thanks to a series of patches from Jonatan Martens.  Wrap javascript control of flash players in myth remote/keyboard commands.

Allows fullscrean playback of Youtube and Vimeo content with no themed borders or widgets, as well as seeking and volume control with the remote or keyboard.  This makes the whole experience much better "on the couch."  We'll look into how many of the other scripts can be retrofitted for this kind of control, patches to update them are very welcome.

Fixes ticket 9301.