Don't create an additional thread for media playback.
After the threading changes prior to the 0.24 release, this extra thread
serves no purpose and is slowing down both startup and exiting of video
playback. It may also have an impact on sluggish keypress responses
reported by several people.

There are various cleanups that can now be made in the TV class.