Internet Content:  Convert blip.tv grabber to accept remote input.
Blip.tv decided to go their own way with their javascript API, and it shows.  It stinks.  Theoretically the player is supposed to update member variables with a required callback, but even though I've implemented it *exactly* es in their documentation, the callback is never triggered.  So, instead, I'm forced to track certain states manually.  It causes some occasional strange behavior but volume, seek, pause, and play at least mostly work.  It's pretty easy to tell the difference between the polish of the Vimeo and Youtube javascript controls and the blip.tv one.