Internet Content:  Serve the BBC iPlayer content from the backend.
Doesn't convert the script to remote control-- there doesn't seem to be a public javascript API, though there might be some info out there/there might be a non-documented API.  If anyone wants to do that part of the work it'd be appreciated and wouldn't take long if someone knwe what they were doing.  This at least prevents the need to have the iplayer script installed on the Frontend.