Blu-ray Playback: Some improvements to HDMV menu image handling.
With some help from Mark Kendall, the images are now in the proper colorspace.  Also a nasty, dirty hack to make sure that the overlay images are all divisible by four, which prevents their being corrupted (but unfortunately will render their eventual placement onscreen imperfect).  I'm following this up with the upstream author of the code this uses and will update the library when a better fix is available.