MythVideo:  Start to work on some of the redundancy/overcomplication issues.
"Mark as watched" appeared not one, but two spots in the menu.  The menu also contained a submenu with that option, as well as an option to mark something as non-browseable.  Since that's not going to be a commonly-invoked toggle, and since browseability is also toggleable from the edit metadata screen, remove the submenu entirely.

Also remove the "Manually enter video ID #" option in the metadata menu-- This can and should be done in the edit metadata menu.