MythVideo:  Only show the "Delete" option from the Video Manager.
Consequently, give the video manager a reason to exist again.  Lots of people have complained that the Delete option is too exposed/available and I am inclined to agree.  Hopefully mitigate some of this by making it accessible only in the manager view, which takes a much more deliberate set of steps to access.