Remove the state change QWaitCondition from the TV class.
Now that the additional TV thread has been removed, this is no longer
necessary and has a large negative impact on playback under certain
conditions. In cases where the frame rate closely matches the display's
refresh rate, the delay caused by the QWaitCondition is large enough to
cause significant 'jitter' and may be responsible for a number of
reported problems in the 0.24 release. The while loop in TV::StartTV is
now effectively the main loop during media playback and needs to be as
simple as possible.

If there are no other problems, I will backport in a few days.