Blu-ray Playback: Remove image padding hack, add mouse click support to BDRB.
I'm pulling the image padding hack for now-- there's circumstantial evidence that it may be an issue with the PNG output code, and that the images themselves may be okay.

Also add a ClickButton function to the BDRingBuffer.  It's not used anywhere, but it does hook into the API correctly, and someone who knows about mouse handling in the player beter than I might have an easier time getting it to work.