Build ffmpeg as mythffmpeg and install
This will allow nuvexport to standardize on using the snapshotted ffmpeg which
is built against the libmythav* libraries, as is mythtv itself.  This will
get rid of the constant changes in ffmpeg command-line arguments once and for
all.  At every ffmpeg resync, we will need to be sure that this is recreated,
and then that nuvexport uses whatever changed command-line arguments that the
new sync uses.