MythMusic: allow the music player to be controlled by sending it MythEvents.
Currently you can start, stop and pause the player, get/set the volume, play a
file, url or track id and get the metadata of the currently playing track.
Other commands will be added as they are needed.

Shortly I will add some code to the web browser widget that will allow the
music player to be controlled using JavaScript.

It would be trivial to update the network control to allow the player to be
controlled using that as well if anyone is interested.