Allows a web page viewed in MythUIWebBrowser to control the music player.
This adds a MusicPlayer java script object that can control the music player.
The main use for this will be in MythNetVision to allow it to add audio only
feeds.

The object name is 'MusicPlayer' and currently has these functions.

Play()             - Starts the music player.
Pause()            - toggles the pause state of the music player.
Stop()             - stops the music player.
SetVolume(int)     - set the volume, takes a number between 0 and 100
GetVolume()        - returns the volume
GetMetadata()      - returns a string containing the current track details.
PlayFile(filename) - starts playing filename which should be a local file.
PlayURL(url)       - starts playing url which can be an http url.
PlayTrack(trackid) - starts playing the track from the database with id trackid.