Remove ALSA mmap access
MMAP is only available for PCI-type audio devices. It isn't available for USB, pulseaudio emulation and other similar devices. Unfortunately, alsa isn't always 100% reliable in determining if mmap direct access is available. So we remove it. ALSA will use mmap access internally if it is available. Mythtv has historically used mmap, can't find a particular reason for it and no-one else seems to use it either