Fixed more replex warnings.
Seems the replex code is compiled into mythtranscode as well.  This catches
those warnings, although at the price of disabling this one:

mpeg2fix.cpp:79: warning: ‘void DEBUGpts(Q3PtrList<MPEG2frame>*)’ defined but not used

As this one is just a function that never gets used, I think we can live with
this.