Change on how we calculate the required ALSA hardware buffer size.
Hopefully 4th time lucky. Instead of calculating what we think the size should be, use the size of the buffer ALSA allocated following our request. If less than requested, use the ratio between what we asked for and what we got, and multiply the current hardware buffer size by this value. Round it up the new nearest 64kB multiple