Theme Chooser/Downloader enhancements
Remove the 'show/hide downloadable themes' button from the menu and
default to always displaying downloadable themes.  This is more
friendly for new users and helps promote the fact that we are pushing
towards downloading themes via the downloader rather than installing
packages of groups of themes which the user may never use.

Cache the downloadable theme list rather than re-retrieving the list
every time we display the screen.

Update the busy dialog popup during screen loading so that the user
is informed as to what is going on rather than just seeing "Loading..."
for the duration.