Add a new MythUDPListener class to libmythui.
To test:- mythtvosd --template=alert --alert_text='Hello world!'

This replaces the old udpnotify class in libmythtv and the new object is
owned by MythMainWindow. The UDP port setting is reinstated and moved
from the Playback screen to the General settings pages.

I've retained the XML functionality to ensure we don't pick up stray UDP
broadcast messages and to make the most of the eXtensibility once we've
settled on a general framework and UI for messaging (e.g. to add timeout
settings).

For the time being, mythtvosd will work as above but I will replace it
with a new command line utility (mythmessage?) with the template
handling removed.