libmythmetadata: Add a BlurayMetadata class.
Class will open and parse a given Blu-ray path and if metadata is present, makes it available, including a list of the available titles, a list of thumbnail paths, title, alternate title, etc.  Includes a toMap function for flexibility.

The hope is that this class can be used to present relevant information about a Blu-ray disc in the drive from the menu, eventually.