Changes to MythSystemUnix to fix when a launched program dies quickly
This hopefully will fix #9421 (first report).

Make background runs not block the foreground

The problem with running in the background in from the menu XML is that it is
blocking the drawing, inputdevs, etc, and then launching in the background.
If in the background, there is no reason I can think of that we need to block
the UI thread.  If the program launched in the background has focus, it should
be receiving the keyboard input.  The only one think I'm wondering is if LIRC
will act oddly.

Move timeout setting to before fork

Don't bother managing a child that failed

If the child quickly failed (i.e. errored on start), there is an apparent
race condition that makes it look like the child timed out, and if it was run
in the background, it's deleted before the manager thread gets it, so we core
dump when we try to remove it.