Remove the PlayerTimer class and run the player from TV::StartTV
This is a further natural (and now obvious) improvement following the
threading changes in the player code. Instead of contriving playback
events to trigger the next frame, just loop inside the main event loop
in TV::StartTV and process other application events as needed.

There are almost certainly other simplifications possible here - notably
in the extensive use of timers; many of which can probably now be
removed.

Should almost certainly be a fix for #9223