Extend MythMainWindow to include a base MythRender object by default.
This is the first step (of many) towards using a single, shared hardware
rendering context for the UI and other media. Testing to date shows that
this is a requirement for OS X when using a native Cocoa Qt build, is
almost certainly needed for OpenGL on windows and should remove various
graphical glitches when using Direct3D on windows.