Consolidate display refresh rate calculations in MythDisplay.
This ensures a reasonable refresh rate is always returned (60Hz) and in
the event that a reasonable rate is not available, uses double the video
frame rate (subject to a maximum of 60Hz). The latter fixes timestretch
playback for displays that have no fixed rate (e.g. macboook) following
the new predictive frame skipping code.