Fix Environment Canada animated maps.
As the support for the "pile of images" animation was removed to simplify the
mythweather code, this script stopped working.  Additionally, it was not
returning any copyright data.  As the requirement for Image::Magick has already
been added for some of the mythnetvision scripts, there's little reason not to
use it in mythweather.  To fix this animation, we are now reading in each
downloaded image in turn with Image::Magick, and making an animated GIF to pass
to mythweather.

Fixes #9467