Add some synchronisation betweem the BD ringbuffer and the player.
This is modelled on the existing DVD synchronisation and should help
with some scene transitions when running in interactive mode.

N.B. This is intended purely as an interim measure until some more
robust code is firmed up that is specific to bluray playback. Please do
not start extending and modifying it - it will probably go away.