Add bluray menu options to the OSD menu.
Adds entries for the top and popup menu, renames the player GoToDVDMenu
to GoToMenu to make it more generic and fixes bringing up the popup menu
when not in a menu already (when using the "MENUTEXT" action). The
latter can probably be removed as I have added a JUMPTOPOPUPMENU action
without a default keybinding.

I will probably also rename the JUMPTODVDROOTMENU and associated actions
but was unsure at this stage of possible complications.