Update the timestamp columns in mythplugins.
Following SHA1: 81f3a7e0586df4ca9c0880480a30f9b3faa1036b , update the
timestamp columns in mythplugins to explicitly specify properties since MySQL
has changed default properties several times since 3.23.  This ensures that
all schemas use the correct properties.

Two of the 3 changes in MythMusic are historical--since the table containing
one of the 3 columns was dropped and a 2nd column was changed to a datetime
field.  The remaining one column in MythMusic and one in MythWeather still
exist, but are almost definitely correct in the vast majority of databases.
This change just ensures that they will remain correct regardless of changes
to MySQL server.