Re-visit the bluray menu OSD
- Move the overlay into its own UI/OSD screen. The bluray menus are not
captions and need separate treatment - the popup menu is, for example,
displayed over the current video stream and hence potentially over
subtitles.
- Clear overlay images that are submitted but already displayed (based
upon size and position) and replace them with the new version.
- Clear the overlay when a NULL image is presented (not NULL overlay)
and use the associated size to prepare the overlay.
- Add some sense checking and memory alignment/allocation fixes in the
callback function.

All told this appears to give far more reliable and consistent results,
though the menu does seem to go missing under certain circumstances.

There is also a strange performance problem when using VDPAU (this was
happening before as well).