Blu-ray Playback: Add a metadata disc library member.
This struct gives us access to the disc title, language, alternate title, disc number, total disc number, and some limited per-title information in some cases.  Going to try to figure out how to use this information to populate the OSD when no metadata for the title/subtitle/etc. is present.