Use --no-color when checking the git branch.
Add --no-color when checking the git branch to fix and issue when
users have "branch = always" in their .gitconfig file.  This was
causing color escape sequences to show up in the branch string
in version.cpp.

Thanks to Bill Meek for tracking down the issue which manifested
itself in the theme chooser not being able to download themes
for some users running master.