Bluray: Improved chapter handling.
libbluray adds an extra chapter marker for the end of the title - so
ignore this for our own purposes - and use the libbluray functionality
for determining the current chapter. This aligns our presentation with
the disc menu structure.

I still intend to convert the chapter seeks to use the libbluray methods
but this requires a little more work.

N.B. Position information can go wrong with 'series/season' discs.
libbluray presents us with a 'super' title that is an amalgamation of
all of the episodes on a disc. Once you've navigated beyond the first
actual title, however, all positioning is relative to the current title.