Move the themeinfo.xml download url under a 'downloadurl' tag.
There is a new downloadinfo tag in the downloadable themeinfo.xml
files which contains information about the downloadable theme.
This patch modifies ThemeInfo to pull the download url from this
new tag instead of from within the detail section as it did
previously.  The themeinfo.xml files on the download site will
retain the old tag as well for backwards compatiblity until at
least the next release.