Don't allow ProgramInfoCache to update the UI on initial load.
Don't trigger a UI update in ProgramInfoCache when we enter the
PlaybackBox and load the recordings list for the first time.  The
list was being displayed even in cases where the user has chosen
to prompt for the initial Recording Group or the current group is
protected by a password.  Delaying the UI update allows the user
to be first prompted for the password or new group and then the
regular UI update code handles displaying the recordings list.

Closes #9209.