Allow a theme xml file to include an additional base file.
The base file is similar to the normal base.xml file and can contain font and
widget definitions that can be inherited from. The theme parser stores the
definitions in the global object store the first time the base file is parsed
and so only needs to parse the file once which can speed up window loading
slightly.

One example of where it will be useful is in MythMusic allowing you to define
all the shared widgets like playback controls in the base file and even to use
the <group> widget to group those widgets into logical groups. Adding them to
the various music views is then a simple matter of adding a <group> that
inherits from the base <group> widget.

The include file is added at the top level of a theme file for example:-

<?xml version="1.0" encoding="utf-8"?>
<!DOCTYPE mythuitheme SYSTEM "http://www.mythtv.org/schema/mythuitheme.dtd">

<mythuitheme>

    <include>music-base.xml</include>

    <window name="searchview">
    .
    .
    .
    </window>

</mythuitheme>