Add a simple "First Time Setup" wizard, but do not enable it.
This adds a setup wizard for a couple of basic settings, audio, and video.  It is not active because it is a work in progress and only one non-core theme is themed for it yet.

The setup wizard walks you through setting up and testing an audio device, setting up and testing playback profiles, and will ultimately permit cloning settings from one frontend to another.  In this way, it will be nearly impossible to get through the setup without being absolutely sure that audio and video playback are functional, and that optionally all the settings are the same on all your systems.

Video samples are not yet finalized, and will eventually be provided by downloading them the first time from a myth services site.