MythFrontend: Add a Grabber Settings window.
Adds a central configuration location to choose metadata sources.  Currently supported are Movies, Television, and Games.  We needed this in order to start using the metadata classes in core MythTV to get artwork and additional information about recorded items.  Currently there is only one option each for Movies and Television, but hopefully that will soon change.  There are two options for MythGame but one of them needs a slight tweak before it will appear in this screen.

This commit does not add the option to the menu themes as I have not completed the default themes (nor have I begun them).