Make the TV and Movie grabber settings global.
We don't want to end up with another round of telling people "Set up a frontend on your backend, then configure them" if the backend is to do any metadata lookups.  I'm willing to accept the limitation of one installation-wide grabber for each type, I think it's simple and understandable, versus a per-host nightmare.

Game metadata grabber remains per host as Game metadata grabbing is only done on the frontend, and only will be for the forseeable future.