Change ThreadPool shutdown to use deleteLater
This is to fix #9500, which has the frontend segfaulting on shutdown.  This
code will hopefully all be overhauled in the awaited webserver rewrite, but
in the mean time, let's have it behaving.

I also moved the order of removing from the deque and marking it for deletion,
as I was toying with different locking methods.