Video Setup Wizard: Stage automatic download of sample files.
Lay the groundwork to pull sample files from services.mythtv.org (or wherever) in the near future when running the video configuration wizard.  The file is sought on the Master Backend's temp storage group, and if it's not found, it's downloaded to that location.  Thus, the download is only ever necessary once in the entire system (and we don't need to distribute 100 MB of samples with the source).